Remove alter strategy
We don't want it (it doesn't create schema that matches up with our production
schema) and it's dangerous, in that it could lead to production tables being
dropped. We don't want that to happen, so let's remove alter from the codebase.

There's one test which relies on the `name` and `age` properties being set at
the top level. These were done in the `alter` strategy. I assume it's safe to
remove because we're not using `'alter'` currently.

The `'alter'` strategy was the only thing that relied on `prompt` so we can
also remove that dependency from the codebase.