Remove findOrCreate, findOrCreateEach
We don't use these functions since they're race-y, and even if we wanted to
implement a find-or-create function, we'd want to a) write it ourselves (e.g.
make it purposefully hard to write) and b) reverse the order, so you attempt
the create first, then find, so we're not vulnerable to races.