Remove afterCreate lifecycle callback
We don't want it, and it complicates the code for create().

FWIW, we use this in exactly one place in the codebase which I just submitted
a pull request to remove. Once that's changed, this is fair game. (The tests
will break if we try to deploy this in the API before the other PR is merged).