Use the original WLValidationError
The previous commit bypassed the WLValidationError entirely, but we rely on
the E_VALIDATION code in a bunch of places. So set it, and copy over other
fields from the Anchor error to the WLValidationError - specifically, we
always set `this.message` to be equal to the Anchor error message, and `rule`,
`actualType`, `expectedType`, and `data`, if they are set on the Anchor error
object. These may or may not be set, dependending on the validation failure.

In addition, avoids the `util.inherits` code, and the weird
defineGetter/defineSetter code. I copied the error construction/inheriting code
from our own Shyp API error constructing code, which seems to be doing just
fine.