Add better validation error messages for types
Upgrades lusitania to v2.2.0, which includes the column name as part of the
rule validation error message. This lets you know which field failed in the
WLValidationError without depending on a huge hacked up `invalidAttributes`
struct that no one could decipher or figure out how to use well.

The error messages aren't perfect and I'm hoping to improve on them soon, but
I think they're an improvement, especially the required, max/min and length
error messages.

Fixes tests which incorrectly encoded the belief that gender is a binary.